+++ install.sh
@@ -141,11 +141,8 @@ sudo ln -sfnv ${HOMEBREW_PATH}/etc/nginx/sites-available/default ${HOMEBREW_PATH
 sudo cp "${HOMEBREW_PATH}/Cellar/nginx/1.23.1/html/index.html /var/www/"
 sudo chown -R $(whoami):staff /var/www
 echo "${yellow}Installing PHP.${txtreset}"
-brew tap homebrew/homebrew-php
-brew install php@7.4
-mkdir -p ~/Library/LaunchAgents
-cp /usr/local/opt/php@7.4/homebrew.mxcl.php@7.4.plist ~/Library/LaunchAgents/
-sudo launchctl load -w ~/Library/LaunchAgents/homebrew.mxcl.php\@7.4.plist 
+brew tap shivammathur/php
+brew install shivammathur/php/php@@7.4
 lsof -Pni4 | grep LISTEN | grep php
 sudo ln -s /System/Volumes/Data/opt/homebrew/etc/php /etc/PHP
 sudo echo "export PATH=\"\$(brew --prefix php@7.4)/bin:\$PATH\"" >> ~/.bash_profile